Cleanup config types.

+++ parachain/pallets/ethereum-beacon-client/src/mock.rs
@@ -82,8 +82,8 @@ pub mod mock_minimal {
 		pub const MaxVoluntaryExitSize: u32 = config::MAX_VOLUNTARY_EXITS as u32;
 		pub const MaxAttestationSize: u32 = config::MAX_ATTESTATIONS as u32;
 		pub const MaxValidatorsPerCommittee: u32 = config::MAX_VALIDATORS_PER_COMMITTEE as u32;
-		pub const MaxSlotsPerHistoricalRoot: u32 = 64u32;
-		pub const MaxFinalizedHeaderSlotArray: u32 = 1000u32;
+		pub const MaxSlotsPerHistoricalRoot: u64 = 64;
+		pub const MaxFinalizedHeaderSlotArray: u32 = 1000;
 		pub const WeakSubjectivityPeriodSeconds: u32 = 97200;
 		pub const ChainForkVersions: ForkVersions = ForkVersions{
 			genesis: Fork {
@@ -196,8 +196,8 @@ pub mod mock_mainnet {
 		pub const MaxVoluntaryExitSize: u32 = config::MAX_VOLUNTARY_EXITS as u32;
 		pub const MaxAttestationSize: u32 = config::MAX_ATTESTATIONS as u32;
 		pub const MaxValidatorsPerCommittee: u32 = config::MAX_VALIDATORS_PER_COMMITTEE as u32;
-		pub const MaxSlotsPerHistoricalRoot: u32 = 8192u32;
-		pub const MaxFinalizedHeaderSlotArray: u32 = 1000u32;
+		pub const MaxSlotsPerHistoricalRoot: u64 = 8192;
+		pub const MaxFinalizedHeaderSlotArray: u32 = 1000;
 		pub const WeakSubjectivityPeriodSeconds: u32 = 97200;
 		pub const ChainForkVersions: ForkVersions = ForkVersions{
 			genesis: Fork {
